+++ src/components/ActQuestion.js
@@ -1,0 +1,9 @@
+const ActQuestion = ({ questionImg }) => {
+  return (
+    <div className="act-question">
+      <img src={questionImg} alt="act matrix question" />
+    </div>
+  );
+};
+
+export default ActQuestion;
