+++ src/components/App.js
@@ -1,0 +1,13 @@
+import Header from "./Header";
+import ActView from "./ActView";
+
+function App() {
+  return (
+    <div className="App">
+      <Header />
+      <ActView />
+    </div>
+  );
+}
+
+export default App;
