+++ src/components/Header.js
@@ -1,0 +1,3 @@
+const Header = () => <h2 className="header">Let's Matrix</h2>;
+
+export default Header;
