+++ src/questions.js
@@ -1,0 +1,15 @@
+import questionOne from "./images/act-matrix-q1.png";
+import questionTwo from "./images/act-matrix-q2.png";
+import questionThree from "./images/act-matrix-q3.png";
+import questionFour from "./images/act-matrix-q4.png";
+import questionFive from "./images/act-matrix-q5.png";
+
+const questions = [
+  questionOne,
+  questionTwo,
+  questionThree,
+  questionFour,
+  questionFive,
+];
+
+export default questions;
